+++ noodles-cram/CHANGELOG.md
@@ -30,6 +30,9 @@
   * cram/codecs/rans_nx16/encode: Disable rANS encoding when the input is
     empty.
 
+  * cram/codecs/rans_nx16/encode: Disable bit packing if the input has more
+    than 16 symbols.
+
   * cram/codecs/rans_nx16/encode/order_1: Fix order of rANS state
     renormalizations.